requirements.txt
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
dj-database-url==2.1.0
Django==5.0.3
django-environ==0.11.2
environ==1.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
python-telegram-bot==21.0.1
sniffio==1.3.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
